major restructuring for single node process
