+++ api/controller/journal.js
