+++ docs/www/api_project.js
@@ -43,7 +43,7 @@ define({
   "apidoc": "0.3.0",
   "generator": {
     "name": "apidoc",
-    "time": "2017-07-08T18:47:17.666Z",
+    "time": "2017-07-16T17:49:55.521Z",
     "url": "http://apidocjs.com",
     "version": "0.17.6"
   }
