+++ package.json
@@ -25,7 +25,7 @@
 	},
 	"scripts": {
 		"test": "mocha api/test/",
-		"docs": "apidoc -i api/routes -o docs/www/",
+		"docs": "apidoc -i api/controller -o docs/www/",
 		"run": "node sugarizer.js"
 	}
 }
